Add plots for clans and correct true characteristics to 0.5 if tau is constant.
