+++ src/topics_causal_inf/generic_ml/simulation.py
@@ -133,7 +133,7 @@ def _simulate_true_clan(
     """Simulate true CLAN for the given DGP."""
     # Check whether dgp.treatment_effect.keywords["tau"] raises error
     if "tau" in dgp.treatment_effect.keywords:  # type: ignore[union-attr]
-        return np.ones((2, dim)) * dgp.treatment_effect.keywords["tau"]  # type: ignore[union-attr]
+        return np.ones((2, dim)) * 0.5  # type: ignore[union-attr]
 
     data = data_wager_athey_2018(
         n_obs=n_obs,
